#11 Move package (b4rpipe → b4r.pipe)
