+++ b4r/__init__.py
@@ -1,0 +1,6 @@
+__all__ = ["pipe"]
+__version__ = "0.1.3"
+
+
+# subpackages
+from . import pipe
