+++ b4r/pipe/__init__.py
@@ -1,0 +1,1 @@
+from b4r.pipe.main import *
