+++ b4r/pipe/godec.py
