--- b4rpipe/__init__.py
@@ -1,8 +1,0 @@
-from b4rpipe.main import *
-
-__copyright__ = "Copyright (C) 2019-2021 B4R team"
-__version__ = "0.1.3"
-__author__ = "astroysmr"
-__author_email__ = "astro.yoshimura@gmail.com"
-__url__ = "https://github.com/LMT-heterodyne-dev/b4rpipe"
-__all__ = []
